# bugtrace/api/main.py
"""
FastAPI Application - Main application with CORS, health check, and routing.

Provides REST API for BugtraceAI scan and report management.

Solves:
- API-09: CORS with explicit origins (not wildcard with credentials)
- API-10: Serve command integration
- INF-04: Health check endpoint

Author: BugtraceAI Team
Date: 2026-01-27
Version: 4.0.0-rc1
"""

import asyncio
import os
import uuid
from contextlib import asynccontextmanager
from typing import Dict, Any

from fastapi import FastAPI, Request, WebSocket, WebSocketDisconnect, Query
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse

from bugtrace.core.config import settings
from bugtrace.core.database import get_db_manager
from bugtrace.api.deps import ScanServiceDep, ReportServiceDep, EventBusDep, get_scan_service
from bugtrace.api.exceptions import register_exception_handlers
from bugtrace.services.event_bus import service_event_bus
from bugtrace.utils.logger import get_logger, set_correlation_id

logger = get_logger("api.main")

# Module-level state for version check (populated at startup, read by /health)
_update_info: Dict[str, Any] = {}

@asynccontextmanager
async def lifespan(app: FastAPI):
    """
    Lifespan context manager for startup and shutdown.

    Startup:
        - Log API server initialization
        - Verify service dependencies

    Shutdown:
        - Log graceful shutdown
        - Services handle their own cleanup
    """
    # Startup
    logger.info("FastAPI application starting up...")
    logger.info(f"API Version: {settings.VERSION}")
    logger.info(f"Environment: {settings.ENV}")
    logger.info(f"CORS Origins: {_get_cors_origins()}")

    # Clean up orphaned scans (RUNNING/PENDING in DB but no process behind them)
    scan_service = get_scan_service()
    orphaned = scan_service.cleanup_orphaned_scans()
    if orphaned:
        logger.info(f"Marked {orphaned} orphaned scan(s) as FAILED on startup")
    recovered_refreshes = scan_service.recover_pending_repeater_refreshes()
    if recovered_refreshes:
        logger.info(f"Resumed {recovered_refreshes} pending Repeater report refresh(es)")

    # Check for updates (non-blocking, silent on failure)
    global _update_info
    try:
        from bugtrace.utils.version_check import check_for_update_async
        result = await check_for_update_async(settings.VERSION)
        if result:
            _update_info = result
            if result.get("update_available"):
                logger.info(f"Update available: {settings.VERSION} -> {result['latest_version']} ({result['release_url']})")
    except Exception:
        pass

    yield

    # Shutdown
    logger.info("FastAPI application shutting down...")


# OpenAPI tag metadata for Swagger documentation
openapi_tags = [
    {
        "name": "scans",
        "description": "Security scan lifecycle management - create, monitor, stop scans",
    },
    {
        "name": "reports",
        "description": "Scan report retrieval in HTML, JSON, and Markdown formats",
    },
    {
        "name": "config",
        "description": "Runtime configuration viewing and updating",
    },
    {
        "name": "health",
        "description": "Server health and readiness monitoring",
    },
    {
        "name": "metrics",
        "description": "Real-time performance metrics - queue depths, CDP reduction, parallelization",
    },
]

# Create FastAPI app
app = FastAPI(
    title="BugTraceAI API",
    description="REST API for security scanning and report management",
    version=settings.VERSION,
    lifespan=lifespan,
    openapi_tags=openapi_tags,
    docs_url="/docs",
    redoc_url="/redoc",
    contact={"name": "BugTraceAI", "url": "https://github.com/BugTraceAI"},
    license_info={"name": "Apache-2.0", "url": "https://www.apache.org/licenses/LICENSE-2.0.html"},
)


def _get_cors_origins() -> list[str]:
    """
    Get CORS allowed origins from BUGTRACE_CORS_ORIGINS env var.

    Example: BUGTRACE_CORS_ORIGINS=http://localhost:5173,http://192.168.1.50:3000

    If not set, allows all origins (permissive — configure for production).
    """
    env_origins = os.getenv("BUGTRACE_CORS_ORIGINS", "")
    if env_origins:
        origins = [o.strip() for o in env_origins.split(",") if o.strip()]
        logger.info(f"CORS origins: {origins}")
        return origins

    logger.warning("BUGTRACE_CORS_ORIGINS not set — allowing all origins. Set it in .env for production.")
    return ["*"]


# Configure CORS middleware
_cors_origins = _get_cors_origins()
app.add_middleware(
    CORSMiddleware,
    allow_origins=_cors_origins,
    allow_credentials=_cors_origins != ["*"],
    allow_methods=["GET", "POST", "PUT", "DELETE", "OPTIONS"],
    allow_headers=["Content-Type", "Authorization", "X-Scan-ID"],
    expose_headers=["X-Scan-Progress"],
    max_age=3600,
)

# Register global exception handlers
# Must be registered AFTER middleware to ensure proper error handling
register_exception_handlers(app)


# Correlation ID middleware — sets correlation_id per request for structured log tracing
@app.middleware("http")
async def correlation_id_middleware(request: Request, call_next):
    """
    Assign a correlation_id for every request.

    Priority:
        1. scan_id from URL path (e.g. /api/scans/42 -> "scan-42")
        2. X-Correlation-ID or X-Request-ID request header
        3. Auto-generated short UUID
    """
    correlation_id = ""

    # 1. Extract scan_id from path if present
    parts = request.url.path.split("/")
    for i, part in enumerate(parts):
        if part == "scans" and i + 1 < len(parts) and parts[i + 1].isdigit():
            correlation_id = f"scan-{parts[i + 1]}"
            break

    # 2. Check request headers
    if not correlation_id:
        correlation_id = (
            request.headers.get("x-correlation-id")
            or request.headers.get("x-request-id")
            or ""
        )

    # 3. Fallback to generated UUID
    if not correlation_id:
        correlation_id = uuid.uuid4().hex[:8]

    set_correlation_id(correlation_id)
    response = await call_next(request)
    response.headers["X-Correlation-ID"] = correlation_id
    return response


# Root endpoint
@app.get("/")
async def root() -> Dict[str, Any]:
    """
    API information endpoint.

    Returns:
        Dictionary with API name, version, and documentation link
    """
    return {
        "name": "BugTraceAI API",
        "version": settings.VERSION,
        "docs": "/docs",
        "health": "/health",
        "ready": "/ready",
    }


# Health check endpoint
@app.get("/health", tags=["health"])
async def health_check(
    scan_service: ScanServiceDep,
    event_bus: EventBusDep,
) -> Dict[str, Any]:
    """
    Health check endpoint for monitoring and deployment orchestration.

    Returns:
        status: Server health status (healthy/degraded/misconfigured)
        version: API version
        docker_available: Whether Docker is available for tool execution
        active_scans: Number of currently running scans
        event_bus_stats: Event bus statistics
        provider: Active provider name
        provider_ready: Whether provider preset file exists
        api_key_configured: Whether the provider's API key is set
        warnings: List of configuration problems (empty = all good)

    Solves INF-04: Health check for container readiness probes
    """
    warnings: list[str] = []

    # Check Docker availability
    docker_available = False
    try:
        import shutil
        docker_available = shutil.which("docker") is not None
    except Exception as e:
        logger.warning(f"Docker check failed: {e}")

    # Check provider preset availability
    provider_ready = False
    api_key_configured = False
    provider_name = settings.PROVIDER
    try:
        providers_dir = settings.BASE_DIR / "bugtrace" / "data" / "providers"
        preset_path = providers_dir / f"{settings.PROVIDER}.json"
        if preset_path.exists():
            provider_ready = True
            import json as _json
            preset_data = _json.loads(preset_path.read_text())
            key_env = preset_data.get("api_key_env", "")
            key_value = os.environ.get(key_env) or getattr(settings, key_env, None)
            api_key_configured = bool(key_value)
            provider_name = preset_data.get("name", settings.PROVIDER)
            if not api_key_configured:
                warnings.append(f"API key not configured for provider '{provider_name}'. Set {key_env} in your .env file.")
        else:
            warnings.append(f"Provider preset '{settings.PROVIDER}' not found. Check bugtrace/data/providers/ directory.")
    except Exception as e:
        logger.warning(f"Provider check failed: {e}")
        warnings.append(f"Provider check error: {e}")

    if not docker_available:
        warnings.append("Docker not available. External tools (Nuclei, SQLMap) will be disabled.")

    # Get active scan count
    active_scans = scan_service.active_scan_count

    # Get event bus stats
    event_stats = event_bus.get_stats()

    # Determine overall status
    if not provider_ready or not api_key_configured:
        status = "misconfigured"  # Engine online but can't run scans
    elif not docker_available:
        status = "degraded"  # Can run API but some tools disabled
    else:
        status = "healthy"

    return {
        "status": status,
        "version": settings.VERSION,
        "docker_available": docker_available,
        "provider": provider_name,
        "provider_ready": provider_ready,
        "api_key_configured": api_key_configured,
        "active_scans": active_scans,
        "event_bus_stats": event_stats,
        "warnings": warnings,
        "update_available": _update_info.get("update_available", False),
        "latest_version": _update_info.get("latest_version"),
    }


# Readiness check endpoint
@app.get("/ready", tags=["health"])
async def readiness_check() -> Dict[str, Any]:
    """
    Readiness check endpoint for deployment orchestration.

    Unlike /health (liveness probe) which indicates the server process is running,
    /ready (readiness probe) verifies that the application can serve requests by
    checking external dependencies like database connectivity.

    Returns:
        ready: Overall readiness status (true only if database is reachable)
        checks:
            database: Whether SQLite database is accessible
            docker_available: Whether Docker is available for scan execution

    Use /health for liveness probes (is the process alive?).
    Use /ready for readiness probes (can it handle requests?).
    """
    # Check database connectivity using DatabaseManager (not direct sqlite3)
    database_ok = False
    try:
        db = get_db_manager()
        with db.get_session() as session:
            from sqlalchemy import text
            session.exec(text("SELECT 1"))
        database_ok = True
    except Exception as e:
        logger.warning(f"Database readiness check failed: {e}")

    # Check Docker availability
    docker_available = False
    try:
        import shutil
        docker_available = shutil.which("docker") is not None
    except Exception as e:
        logger.warning(f"Docker readiness check failed: {e}")

    # Overall readiness: database must be reachable (Docker is optional)
    ready = database_ok

    return {
        "ready": ready,
        "checks": {
            "database": database_ok,
            "docker_available": docker_available,
        },
    }


# Router includes
from bugtrace.api.routes.scans import router as scans_router
from bugtrace.api.routes.reports import router as reports_router
from bugtrace.api.routes.config import router as config_router
from bugtrace.api.routes.metrics import router as metrics_router
from bugtrace.api.routes.websocket import router as websocket_router
from bugtrace.api.routes.providers import router as providers_router
from bugtrace.api.routes.oob import router as oob_router
from bugtrace.api.routes.model_eval import router as model_eval_router

app.include_router(scans_router, prefix="/api", tags=["scans"])
app.include_router(reports_router, prefix="/api", tags=["reports"])
app.include_router(config_router, prefix="/api", tags=["config"])
app.include_router(providers_router, prefix="/api", tags=["providers"])
app.include_router(metrics_router, prefix="/api", tags=["metrics"])
app.include_router(websocket_router, prefix="/api", tags=["websocket"])
app.include_router(model_eval_router, prefix="/api", tags=["model-eval"])
app.include_router(oob_router, prefix="/api", tags=["oob"])


# WebSocket endpoint for real-time scan event streaming
@app.websocket("/ws/scans/{scan_id}")
async def websocket_scan_stream(
    websocket: WebSocket,
    scan_id: int,
    last_seq: int = Query(default=0, description="Last received sequence number for reconnection"),
):
    """
    WebSocket endpoint for streaming real-time scan events.

    Args:
        websocket: WebSocket connection
        scan_id: Scan ID to stream events for
        last_seq: Last received sequence number (for reconnection, default 0)

    Behavior:
        1. Accept WebSocket connection
        2. Validate scan_id exists
        3. If last_seq > 0 (reconnection): Send missed events first
        4. Stream live events using ServiceEventBus.stream()
        5. Handle WebSocketDisconnect gracefully
        6. Close cleanly on scan completion (scan_complete or error events)

    Reconnection support:
        Client can reconnect with ?last_seq=N to receive only missed events
        since sequence number N.

    Solves WS-02: Real-time event streaming with reconnection support
    """
    await websocket.accept()
    logger.info(f"WebSocket client connected to scan {scan_id} (last_seq={last_seq})")

    try:
        if not await _ws_validate_scan(websocket, scan_id):
            return
        highest_sent = await _ws_send_missed_events(websocket, scan_id, last_seq)
        await _ws_stream_live_events(websocket, scan_id, highest_sent)
        await websocket.close(code=1000, reason="Scan complete")
    except WebSocketDisconnect:
        logger.info(f"WebSocket client disconnected from scan {scan_id}")
    except Exception as e:
        await _ws_handle_error(websocket, scan_id, e)


async def _ws_validate_scan(websocket: WebSocket, scan_id: int) -> bool:
    """Validate scan exists before streaming."""
    scan_service = get_scan_service()
    try:
        await scan_service.get_scan_status(scan_id)
        return True
    except Exception as e:
        logger.warning(f"Invalid scan_id {scan_id}: {e}")
        await websocket.close(code=1008, reason=f"Invalid scan_id: {scan_id}")
        return False


async def _ws_send_missed_events(websocket: WebSocket, scan_id: int, last_seq: int) -> int:
    """Send missed events for reconnection."""
    if last_seq == 0:
        return 0

    logger.info(f"Reconnection detected for scan {scan_id}, replaying events since seq {last_seq}")
    missed = service_event_bus.get_history(scan_id, since_seq=last_seq)
    for event in missed:
        await websocket.send_json(event)
    return missed[-1]["seq"] if missed else last_seq


async def _ws_stream_live_events(websocket: WebSocket, scan_id: int, highest_sent: int):
    """Stream live events with deduplication and disconnect detection."""
    stream_task = asyncio.create_task(
        _ws_drain_event_stream(websocket, scan_id, highest_sent)
    )

    try:
        while not stream_task.done():
            receive_task = asyncio.create_task(websocket.receive_text())
            done, pending = await asyncio.wait(
                {stream_task, receive_task},
                return_when=asyncio.FIRST_COMPLETED,
            )

            if receive_task in done:
                try:
                    await receive_task
                except WebSocketDisconnect:
                    logger.info(f"WebSocket client disconnected from scan {scan_id}")
                    break
                except Exception:
                    logger.info(f"WebSocket receive ended for scan {scan_id}")
                    break
                continue

            for task in pending:
                task.cancel()
            await stream_task
            break
    finally:
        if not stream_task.done():
            stream_task.cancel()
            try:
                await stream_task
            except asyncio.CancelledError:
                pass


async def _ws_drain_event_stream(websocket: WebSocket, scan_id: int, highest_sent: int):
    """Drain service event bus events into the websocket.

    Skip/terminal rules: pure owner ``ws_envelope_policy`` (P5-WS-1).
    """
    from bugtrace.api.ws_envelope_policy import (
        is_terminal_event_type,
        should_skip_stream_event,
    )

    async for event in service_event_bus.stream(scan_id):
        if should_skip_stream_event(event, highest_sent):
            continue

        await websocket.send_json(event)
        highest_sent = event.get("seq", highest_sent)

        if is_terminal_event_type(event.get("event_type")):
            logger.info(
                f"Scan {scan_id} completed with event_type={event.get('event_type')}, "
                "closing WebSocket"
            )
            break


async def _ws_handle_error(websocket: WebSocket, scan_id: int, error: Exception):
    """Handle WebSocket errors gracefully."""
    logger.error(f"WebSocket error for scan {scan_id}: {error}", exc_info=True)
    try:
        await websocket.close(code=1011, reason="Internal server error")
    except Exception:
        # WebSocket already closed or connection lost - ignore
        pass