# openapi.yaml
openapi: 3.1.0
info:
  title: BugTraceAI CLI API
  description: |
    REST API for the BugTraceAI autonomous security scanning platform.

    BugTraceAI CLI is a headless, AI-powered vulnerability scanner that discovers and validates
    security issues using specialized agents (XSS, SQLi, SSRF, IDOR, LFI, RCE, XXE, JWT, and more).
    This API provides full scan lifecycle management, real-time WebSocket streaming, report generation,
    and runtime configuration.

    **Local beta deployment only.** No authentication is required because access is assumed to be
    limited to the local machine or a trusted LAN. Do not expose the API directly to the Internet.
    ModelLab uses the configured OpenRouter API key and may incur provider costs.
    WebSocket endpoints are available for real-time scan events, findings, and ModelLab progress.

    **Note:** The following endpoints are documented in code but not yet in this spec:
    - POST /api/auth/validate (YAML auth configuration validation)
    - PATCH /api/scans/{scan_id}/auth-config (Update scan authentication)
    - POST /api/tools/totp-verify (TOTP/2FA verification)
  version: 4.0.0-rc1
  contact:
    name: BugTraceAI
    url: https://bugtraceai.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://{host}:8000
    description: Custom deployment
    variables:
      host:
...
paths:
...
  /ws/scans/{scan_id}:
    get:
      tags: [websocket]
      summary: Scan events (WebSocket)
      description: |
        **WebSocket endpoint** — stream real-time events for a specific scan.

        Connect with `?last_seq=N` to replay missed events on reconnection.

        **Event types:** `progress_update`, `phase_update`, `finding_discovered`, `log`, `scan_complete`, `error`
      operationId: scanWebSocket
      parameters:
        - $ref: "#/components/parameters/ScanId"
        - name: last_seq
          in: query
          schema:
            type: integer
            default: 0
          description: Last received sequence number for reconnection replay
      responses:
        "101":
          description: WebSocket upgrade
        "400":
          description: Invalid scan ID
...
components:
  parameters:
    ScanId:
      name: scan_id
      in: path
      required: true
      schema:
        type: integer
      description: Unique scan identifier
...